Adjustments to Schema Evolution New Attribute Reads (#2484)
* call generate_uri() on attribute drop for schema evolution

* Array Schema needs checked and load per fragment

The array schema is used to determine some details in the fragment
footers. While this is not ideal, it is how the current code behaves. As
such we need to load the schema name first and then set the proper array
schema on each fragment.

We also need to make sure to generate unique URIs on schema evolution
invocation.

* Fix support for returning fill-in value with new attributes

This properly supports new attributes added from schema evolution
returning the fill-in value as the default when older fragments are
included in a query. Minor adjustments are made to the read algorithm to
support the cases where an attribute or dimension's tile does not exist
due to the attribute not existing in the older fragment.

* Add schema evolution example